[test #58] KPI 취소 승인/반려 테스트 코드 구현
