+++ momentum-dao-be/src/main/java/com/dao/momentum/evaluation/kpi/command/application/controller/KpiCommandController.java
@@ -3,7 +3,6 @@ package com.dao.momentum.evaluation.kpi.command.application.controller;
 import com.dao.momentum.common.dto.ApiResponse;
 import com.dao.momentum.evaluation.kpi.command.application.dto.request.*;
 import com.dao.momentum.evaluation.kpi.command.application.dto.response.CancelKpiResponse;
-import com.dao.momentum.evaluation.kpi.command.application.dto.response.KpiApprovalResponse;
 import com.dao.momentum.evaluation.kpi.command.application.dto.response.KpiCreateResponse;
 import com.dao.momentum.evaluation.kpi.command.application.service.KpiCommandService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -17,7 +16,7 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/kpis")
 @RequiredArgsConstructor
-@Tag(name = "KPI", description = "KPI 생성, 취소 요청, 승인/반려 처리 API")
+@Tag(name = "KPI", description = "KPI 생성, 취소 요청 API")
 public class KpiCommandController {
 
     private final KpiCommandService kpiCommandService;
@@ -44,27 +43,4 @@ public class KpiCommandController {
         return ApiResponse.success(response);
     }
 
-    @PatchMapping("/{kpiId}/approval")
-    @Operation(summary = "KPI 작성 승인/반려", description = "팀장이 KPI 작성 요청을 승인하거나 반려 처리합니다. 반려 시 사유 입력이 필요합니다.")
-    public ApiResponse<KpiApprovalResponse> approveKpi(
-            @AuthenticationPrincipal UserDetails userDetails,
-            @PathVariable Long kpiId,
-            @RequestBody @Valid KpiApprovalRequest request
-    ) {
-        Long managerId = Long.parseLong(userDetails.getUsername()); // 인증된 사용자 ID
-        KpiApprovalResponse response = kpiCommandService.approveKpi(managerId, kpiId, request);
-        return ApiResponse.success(response);
-    }
-
-    @PatchMapping("/{kpiId}/cancel/approval")
-    @Operation(summary = "KPI 취소 승인/반려", description = "팀장이 KPI 취소 요청에 대해 승인 또는 반려 처리합니다.")
-    public ApiResponse<KpiApprovalResponse> approveCancelKpi(
-            @AuthenticationPrincipal UserDetails userDetails,
-            @PathVariable Long kpiId,
-            @RequestBody @Valid KpiCancelApprovalRequest request
-    ) {
-        Long managerId = Long.parseLong(userDetails.getUsername());
-        KpiApprovalResponse response = kpiCommandService.approveCancelRequest(managerId, kpiId, request);
-        return ApiResponse.success(response);
-    }
 }
