+++ momentum-dao-be/src/main/java/com/dao/momentum/evaluation/kpi/command/application/service/KpiCommandService.java
@@ -1,10 +1,7 @@
 package com.dao.momentum.evaluation.kpi.command.application.service;
 
-import com.dao.momentum.evaluation.kpi.command.application.dto.request.KpiApprovalRequest;
-import com.dao.momentum.evaluation.kpi.command.application.dto.request.KpiCancelApprovalRequest;
 import com.dao.momentum.evaluation.kpi.command.application.dto.request.KpiCreateDTO;
 import com.dao.momentum.evaluation.kpi.command.application.dto.response.CancelKpiResponse;
-import com.dao.momentum.evaluation.kpi.command.application.dto.response.KpiApprovalResponse;
 import com.dao.momentum.evaluation.kpi.command.application.dto.response.KpiCreateResponse;
 
 public interface KpiCommandService {
@@ -14,11 +11,5 @@ public interface KpiCommandService {
 
     // KPI 취소 요청
     CancelKpiResponse cancelKpi(Long empId, Long kpiId, String reason);
-
-    // KPI 승인/반려 처리
-    KpiApprovalResponse approveKpi(Long managerId, Long kpiId, KpiApprovalRequest request);
-
-    // KPI 취소 승인/반려 처리
-    KpiApprovalResponse approveCancelRequest(Long managerId, Long kpiId, KpiCancelApprovalRequest request);
 
 }
