+++ momentum-dao-be/src/test/java/com/dao/momentum/evaluation/kpi/command/application/service/KpiCommandServiceImplTest.java
@@ -159,77 +159,4 @@ class KpiCommandServiceImplTest {
         assertEquals(ErrorCode.KPI_INVALID_STATUS, ex.getErrorCode());
     }
 
-    @DisplayName("KPI 승인/반려 처리")
-    @Nested
-    class ApproveKpiTests {
-
-        private final Long managerId = 10L;
-        private final Long kpiId = 200L;
-
-        private Kpi mockKpi(Status status, UseStatus isDeleted) {
-            return Kpi.builder()
-                    .kpiId(kpiId)
-                    .empId(1L)
-                    .statusId(status.getId())
-                    .isDeleted(isDeleted)
-                    .build();
-        }
-
-        @Test
-        @DisplayName("작성 승인 성공")
-        void approveKpi_success() {
-            Kpi kpi = mockKpi(Status.PENDING, UseStatus.N);
-            when(kpiRepository.findById(kpiId)).thenReturn(Optional.of(kpi));
-
-            KpiApprovalRequest req = new KpiApprovalRequest(true, null);
-            KpiApprovalResponse res = service.approveKpi(managerId, kpiId, req);
-
-            assertThat(res.getKpiId()).isEqualTo(kpiId);
-            assertThat(res.getStatus()).isEqualTo(Status.ACCEPTED.name());
-            assertThat(res.getMessage()).contains("승인");
-        }
-
-        @Test
-        @DisplayName("작성 반려 성공 (사유 입력됨)")
-        void rejectKpi_success_withReason() {
-            Kpi kpi = mockKpi(Status.PENDING, UseStatus.N);
-            when(kpiRepository.findById(kpiId)).thenReturn(Optional.of(kpi));
-
-            KpiApprovalRequest req = new KpiApprovalRequest(false, "목표 불명확");
-            KpiApprovalResponse res = service.approveKpi(managerId, kpiId, req);
-
-            assertThat(res.getKpiId()).isEqualTo(kpiId);
-            assertThat(res.getStatus()).isEqualTo(Status.REJECTED.name());
-            assertThat(res.getMessage()).contains("반려");
-        }
-
-        @Test
-        @DisplayName("작성 반려 실패 (사유 없음)")
-        void rejectKpi_fail_withoutReason() {
-            Kpi kpi = mockKpi(Status.PENDING, UseStatus.N);
-            when(kpiRepository.findById(kpiId)).thenReturn(Optional.of(kpi));
-
-            KpiApprovalRequest req = new KpiApprovalRequest(false, null);
-
-            KpiException ex = assertThrows(KpiException.class, () ->
-                    service.approveKpi(managerId, kpiId, req)
-            );
-            assertThat(ex.getErrorCode()).isEqualTo(ErrorCode.KPI_REJECTION_REASON_REQUIRED);
-        }
-
-        @Test
-        @DisplayName("이미 처리된 KPI는 승인/반려 불가")
-        void rejectKpi_fail_alreadyProcessed() {
-            Kpi kpi = mockKpi(Status.ACCEPTED, UseStatus.N);
-            when(kpiRepository.findById(kpiId)).thenReturn(Optional.of(kpi));
-
-            KpiApprovalRequest req = new KpiApprovalRequest(true, null);
-
-            KpiException ex = assertThrows(KpiException.class, () ->
-                    service.approveKpi(managerId, kpiId, req)
-            );
-            assertThat(ex.getErrorCode()).isEqualTo(ErrorCode.KPI_ALREADY_PROCESSED);
-        }
-    }
-
 }
